requirements.txt
PyYAML>=6.0
pydantic>=2.7
numpy>=1.26
pandas>=2.2
torch>=2.2
fastapi>=0.111.0
uvicorn>=0.30.0
Jinja2>=3.1
python-multipart>=0.0.9
